deps: Bump sdk to v3, detach program from local interface
#### Problem

The sdk v3 crates are out, but the interface is still on v2.

#### Summary of changes

Bump all the dependencies, and have the program use the interface from
crates.io.

The only change concerns the program error.

While going through this, I noticed that we weren't enabling
serde-traits for testing, so enable that too.